:art: :truck: move fixtures to own module
