+++ app/tests/e2e/issuer/ld_proof/test_ld_proof_did_key_bbs.py
@@ -15,7 +15,7 @@ from app.models.issuer import SendCredential
 from app.routes.connections import router as con_router
 from app.routes.issuer import router as issuer_router
 from app.routes.oob import router as oob_router
-from app.tests.util.member_connections import FaberAliceConnect
+from app.tests.util.connections import FaberAliceConnect
 from app.tests.util.trust_registry import DidKey
 from app.tests.util.webhooks import check_webhook_state
 from shared import RichAsyncClient
